✅ 해결된 주요 충돌:
- 프론트엔드: App.js, CompanyDashboard.jsx, Careers.jsx
- 백엔드: pom.xml, 각종 Controller/Service/Repository 파일들

🔧 새로 추가된 팀 기능들:
- 면접 분석 시스템 (interview_analysis)
- 면접 질문 생성 (interview_questions)
- 비밀번호 재설정 기능
- 개선된 개인 대시보드 구조
- 새로운 모달 컴포넌트들
- 면접 평가 시스템

💡 유지된 개인 작업:
- 채용 페이지 개선 및 페이지네이션
- 기업 대시보드 기능 강화
- 회사 소개 페이지 업데이트
- 포트폴리오 관련 기능들

+++ backend/src/main/java/com/zoop/backend/controller/InterviewController.java
@@ -32,7 +32,7 @@ public class InterviewController {
     public ResponseEntity<?> getInterviewById(@PathVariable Integer id) {
         try {
             System.out.println("면접 정보 요청 ID: " + id);
-            InterviewScheduleResponseDto interview = aiInterviewScheduleService.getInterviewSchedule(id);
+            InterviewScheduleResponseDto interview = aiInterviewScheduleService.getInterviewSchedule(id.longValue());
             System.out.println("조회된 면접 정보: " + interview);
             return ResponseEntity.ok(interview);
         } catch (Exception e) {
@@ -49,7 +49,7 @@ public class InterviewController {
             @RequestParam Integer candidateId) {
         try {
             System.out.println("면접 정보 요청 - postId: " + postId + ", candidateId: " + candidateId);
-            InterviewScheduleResponseDto interview = aiInterviewScheduleService.getInterviewByPostIdAndCandidateId(postId, candidateId);
+            InterviewScheduleResponseDto interview = aiInterviewScheduleService.getInterviewByPostIdAndCandidateId(postId.longValue(), candidateId.longValue());
             System.out.println("조회된 면접 정보 (by post/candidate): " + interview);
             return ResponseEntity.ok(interview);
         } catch (Exception e) {
@@ -59,7 +59,6 @@ public class InterviewController {
                 .body(new ErrorResponse("해당 공고에 대한 면접 일정을 찾을 수 없습니다: " + e.getMessage()));
         }
     }
-
 }
